requirements.txt
python-dateutil
requests
beautifulsoup4
tzlocal
pydantic
